Add files via upload
Regression.csv, is the file on which we perform data analysis for the first part. 
MyData.csv is the file through we identify top 30% customers
I have tried to simplify the solution as much as possible. Also, the problem statement is easy to understand 
Lastly , I am a learner like most of you, so I am open to feedback.

P.S= Credits to my friend Pawan for his mentoring

Cheers,
Happy Learning